# docs/storage-api.openapi.yaml
openapi: 3.1.0
info:
  title: Deniz Cloud Storage API
  version: 1.0.0
  description: |
    Cloud storage API for Deniz Cloud self-hosted home server.
    Provides file management (upload, download, rename, move, delete),
    folder CRUD, and resumable uploads via the TUS protocol.

    **Base URL:** `https://storage.denizlg24.com`

    ## Authentication

    All endpoints under `/api/uploads/*`, `/api/folders/*`, and `/api/files/*` require authentication.
    Auth endpoints (`/api/auth/me`, `/api/auth/logout`) also require authentication; login is public.

    Two authentication methods are supported:

    - **Bearer Token (JWT):** Obtained via `POST /api/auth/login`. Pass as `Authorization: Bearer <token>`.
    - **API Key:** Pass as `X-API-Key: <key>` header.

    ## Access Control

    - **Private folders** (`/{userId}/...`): Only the owning user can read/write.
    - **Shared folders** (`/shared/...`): All authenticated users can read. Only the owner or a superuser can modify/delete.

    ## Storage Tiering

    Files are stored on SSD (hot) or HDD (cold) based on disk usage watermarks.
    SSD files preserve directory structure; HDD files use a flat layout keyed by file ID.

    ## TUS Resumable Uploads

    File uploads use the [TUS protocol](https://tus.io/) v1.0.0 with `creation` and `termination` extensions.
    Upload sessions expire after 24 hours. Expired sessions are cleaned up automatically.

    ## Error Format

    All errors follow a consistent shape:
...
    {
...
servers:
...
paths:
...
  /api/uploads:
    options:
      operationId: tusOptions
      summary: TUS protocol discovery
      description: |
        Returns TUS server capabilities. No request body or authentication required
        (auth middleware applies to all `/api/uploads/*`, but OPTIONS is handled first).

        The response advertises supported TUS version, extensions, and the
        `Tus-Resumable` header that all subsequent requests must include.
      tags: [Uploads]
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      responses:
        "204":
          description: TUS capabilities
          headers:
            Tus-Version:
              schema:
                type: string
                const: 1.0.0
              description: Supported TUS protocol versions
            Tus-Extension:
              schema:
                type: string
                const: creation,termination
              description: Supported TUS extensions
            Tus-Resumable:
              schema:
                type: string
                const: 1.0.0
              description: TUS protocol version used by this server
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      operationId: createUpload
      summary: Create an upload session
      description: |
        Initiates a new TUS upload session. The client declares the total file size via
        `Upload-Length` and provides file metadata (filename, target folder, optional MIME type)
        via `Upload-Metadata`.

        **Metadata format:** Comma-separated `key base64value` pairs.

        **Required metadata keys:**
        - `filename` — file name (will be normalized to snake_case, extension preserved)
        - `targetFolder` — virtual path of the target folder (e.g., `/{userId}/documents`)

        **Optional metadata keys:**
        - `filetype` — MIME type (e.g., `text/plain`)

        **Upload session lifetime:** 24 hours from creation. Expired sessions are cleaned up automatically.

        On success, the `Location` header contains the upload URL for subsequent PATCH/HEAD/DELETE requests.
      tags: [Uploads]
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: Upload-Length
          in: header
          required: true
          schema:
            type: integer
            minimum: 0
          description: Total file size in bytes
        - name: Upload-Metadata
          in: header
          required: true
          schema:
            type: string
          description: |
            TUS metadata header. Comma-separated `key base64value` pairs.
            Required keys: `filename`, `targetFolder`. Optional: `filetype`.
          example: filename aGVsbG8udHh0,filetype dGV4dC9wbGFpbg==,targetFolder L3VzZXItaWQvZG9jcw==
        - name: Tus-Resumable
          in: header
          required: true
          schema:
            type: string
            const: 1.0.0
          description: TUS protocol version
      responses:
        "201":
          description: Upload session created
          headers:
            Location:
              schema:
                type: string
              description: URL for subsequent upload operations (e.g., `/api/uploads/{id}`)
              example: /api/uploads/550e8400-e29b-41d4-a716-446655440000
            Upload-Offset:
              schema:
                type: integer
                const: 0
              description: Current upload offset (always 0 for new uploads)
            Tus-Resumable:
              schema:
                type: string
                const: 1.0.0
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                missingLength:
                  summary: Missing Upload-Length header
                  value:
                    error:
                      code: MISSING_UPLOAD_LENGTH
                      message: Upload-Length header is required
                invalidLength:
                  summary: Invalid Upload-Length value
                  value:
                    error:
                      code: INVALID_UPLOAD_LENGTH
                      message: Upload-Length must be a non-negative integer
                missingMetadata:
                  summary: Missing Upload-Metadata header
                  value:
                    error:
                      code: MISSING_METADATA
                      message: Upload-Metadata header is required
                missingFilename:
                  summary: Missing filename in metadata
                  value:
                    error:
                      code: MISSING_FILENAME
                      message: filename is required in Upload-Metadata
                missingTargetFolder:
                  summary: Missing targetFolder in metadata
                  value:
                    error:
                      code: MISSING_TARGET_FOLDER
                      message: targetFolder is required in Upload-Metadata
                invalidPath:
                  summary: Invalid target folder path
                  value:
                    error:
                      code: INVALID_PATH
                      message: Path must start with /
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Access denied to target folder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: ACCESS_DENIED
                  message: You do not have access to this folder
        "404":
          description: Target folder not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: FOLDER_NOT_FOUND
                  message: Target folder does not exist
        "409":
          description: File already exists at target path
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: FILE_EXISTS
                  message: A file already exists at this path
  /api/uploads/{id}:
    head:
      operationId: getUploadStatus
      summary: Query upload progress
      description: |
        Returns the current offset (bytes received) and total size for an in-progress upload.
        Used by TUS clients to resume interrupted uploads.

        Returns `404` if the upload does not exist or does not belong to the authenticated user.
        Returns `410` if the upload is no longer in progress (completed, expired, or cancelled).
      tags: [Uploads]
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/UploadId"
      responses:
        "200":
          description: Upload is in progress
          headers:
            Upload-Offset:
              schema:
                type: integer
                minimum: 0
              description: Bytes received so far
            Upload-Length:
              schema:
                type: integer
                minimum: 0
              description: Total declared file size in bytes
            Cache-Control:
              schema:
                type: string
                const: no-store
            Tus-Resumable:
              schema:
                type: string
                const: 1.0.0
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Upload not found or does not belong to user
        "410":
          description: Upload is no longer in progress (completed, expired, or cancelled)

    patch:
      operationId: uploadChunk
      summary: Upload a file chunk
      description: |
        Appends a chunk of binary data to an in-progress upload. The `Upload-Offset` header
        must match the server's current offset exactly (no out-of-order or duplicate chunks).

        **Content-Type** must be `application/offset+octet-stream`.

        When the final byte is received (`bytesReceived === Upload-Length`), the upload is
        automatically finalized:
        1. SHA-256 checksum is computed
        2. Storage tier is determined (SSD if disk usage < watermark, else HDD)
        3. File is moved from temp to final storage location
        4. File record is created in the database
        5. Upload status is set to `completed`
      tags: [Uploads]
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/UploadId"
        - name: Upload-Offset
          in: header
          required: true
          schema:
            type: integer
            minimum: 0
          description: Current byte offset in the file (must match server state)
        - name: Content-Type
          in: header
          required: true
          schema:
            type: string
            const: application/offset+octet-stream
          description: Must be `application/offset+octet-stream`
      requestBody:
        required: true
        content:
          application/offset+octet-stream:
            schema:
              type: string
              format: binary
              description: Raw binary chunk data
      responses:
        "204":
          description: Chunk accepted
          headers:
            Upload-Offset:
              schema:
                type: integer
                minimum: 0
              description: New byte offset after this chunk
            Tus-Resumable:
              schema:
                type: string
                const: 1.0.0
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                missingOffset:
                  summary: Missing Upload-Offset header
                  value:
                    error:
                      code: MISSING_OFFSET
                      message: Upload-Offset header is required
                invalidOffset:
                  summary: Invalid Upload-Offset value
                  value:
                    error:
                      code: INVALID_OFFSET
                      message: Upload-Offset must be a non-negative integer
                emptyBody:
                  summary: No chunk data in body
                  value:
                    error:
                      code: EMPTY_BODY
                      message: Request body is empty
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Upload not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: UPLOAD_NOT_FOUND
                  message: Upload not found
        "409":
          description: Offset mismatch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: OFFSET_MISMATCH
                  message: "Expected offset 1024, got 0"
        "410":
          description: Upload is no longer in progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                finished:
                  summary: Upload already completed or cancelled
                  value:
                    error:
                      code: UPLOAD_FINISHED
                      message: Upload is no longer in progress
                expired:
                  summary: Upload session expired
                  value:
                    error:
                      code: UPLOAD_EXPIRED
                      message: Upload has expired
        "413":
          description: Upload exceeded declared size
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: SIZE_EXCEEDED
                  message: Upload exceeded declared size
        "415":
          description: Wrong Content-Type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: INVALID_CONTENT_TYPE
                  message: Content-Type must be application/offset+octet-stream

    delete:
      operationId: cancelUpload
      summary: Cancel an upload (TUS termination)
      description: |
        Cancels an upload session and deletes the temporary file from disk.
        The upload record is permanently removed from the database.
      tags: [Uploads]
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/UploadId"
      responses:
        "204":
          description: Upload cancelled and temp file deleted
          headers:
            Tus-Resumable:
              schema:
                type: string
                const: 1.0.0
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Upload not found or does not belong to user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: UPLOAD_NOT_FOUND
                  message: Upload not found

  # ──────────────────────────────────────────────
  # Folders
  # ──────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Session JWT obtained via `POST /api/auth/login`.
        Tokens expire after 24 hours by default.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key for programmatic access. Created via the admin panel.
        Pass the raw key value (not hashed).
  parameters:
    UploadId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Upload session UUID (from the `Location` header of the creation response)
...
  responses:
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            noAuth:
              summary: No authentication provided
              value:
                error:
                  code: UNAUTHORIZED
                  message: Authentication required
            tokenInvalid:
              summary: JWT is invalid or expired
              value:
                error:
                  code: TOKEN_INVALID
                  message: Invalid or expired token
            sessionInvalid:
              summary: Session revoked or expired in DB
              value:
                error:
                  code: SESSION_INVALID
                  message: Session expired or revoked
            apiKeyInvalid:
              summary: API key not found
              value:
                error:
                  code: INVALID_API_KEY
                  message: Invalid API key
            apiKeyExpired:
              summary: API key past expiry
              value:
                error:
                  code: API_KEY_EXPIRED
                  message: API key expired
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
...